fix: change metrics name
